Move console stuff into own namespace
