+++ bin/stakx
@@ -3,19 +3,11 @@
 
 require __DIR__ . '/../vendor/autoload.php';
 
-use allejo\stakx\Application;
-use allejo\stakx\Command\BuildCommand;
-use Symfony\Component\Console\CommandLoader\FactoryCommandLoader;
-
-$commandLoader = new FactoryCommandLoader([
-    'build' => function () { return new BuildCommand(); }
-]);
-
-$application = new Application('stakx', '@package_version@');
-$application->setCommandLoader($commandLoader);
+use allejo\stakx\Console\Application;
 
 try
 {
+    $application = new Application('stakx', '@package_version@');
     $application->run();
 }
 catch (\Exception $e)
