--- src/allejo/stakx/Command/CommandInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-/**
- * @copyright 2018 Vladimir Jimenez
- * @license   https://github.com/stakx-io/stakx/blob/master/LICENSE.md MIT
- */
-
-namespace allejo\stakx\Command;
-
-interface CommandInterface
-{
-    const NAME = 'stakx.command';
-}
